기지국 설치
